@@ -1,6 +1,6 @@
 {
   "name": "stream-chat-react-native-core",
-  "version": "2.0.0-beta.1",
+  "version": "2.0.0",
   "author": {
     "company": "Stream.io Inc",
     "name": "Stream.io Inc"
@@ -88,12 +88,12 @@
     }
   },
   "dependencies": {
-    "@babel/runtime": "7.12.1",
+    "@babel/runtime": "7.12.5",
     "@stream-io/react-native-simple-markdown": "1.2.1",
     "anchorme": "^1.1.2",
-    "dayjs": "1.9.4",
-    "file-loader": "6.1.1",
-    "i18next": "19.8.2",
+    "dayjs": "1.9.5",
+    "file-loader": "6.2.0",
+    "i18next": "19.8.3",
     "lodash-es": "4.17.15",
     "metro-react-native-babel-preset": "^0.55.0",
     "mime-types": "2.1.27",
@@ -101,8 +101,9 @@
     "react-art": "^16.8.6",
     "react-native-actionsheet": "2.4.2",
     "react-native-image-zoom-viewer": "3.0.1",
+    "react-native-markdown-package": "1.8.1",
     "seamless-immutable": "7.1.4",
-    "stream-chat": "2.7.0",
+    "stream-chat": "2.7.4",
     "styled-components": "5.2.0",
     "uuid": "8.3.1"
   },
@@ -130,24 +131,24 @@
     "@testing-library/react-native": "7.1.0",
     "@types/eslint": "7.2.4",
     "@types/jest": "26.0.15",
-    "@types/lodash": "4.14.162",
+    "@types/lodash": "4.14.165",
     "@types/mime-types": "2.1.0",
-    "@types/react": "16.9.53",
-    "@types/react-native": "0.63.30",
+    "@types/react": "16.9.56",
+    "@types/react-native": "0.63.33",
     "@types/react-native-actionsheet": "2.4.2",
     "@types/react-test-renderer": "16.9.3",
     "@types/styled-components": "5.1.4",
-    "@typescript-eslint/eslint-plugin": "4.5.0",
-    "@typescript-eslint/parser": "4.5.0",
+    "@typescript-eslint/eslint-plugin": "4.6.1",
+    "@typescript-eslint/parser": "4.6.1",
     "babel-eslint": "10.1.0",
-    "babel-jest": "26.6.1",
+    "babel-jest": "26.6.3",
     "babel-loader": "8.1.0",
     "babel-plugin-i18next-extract": "0.8.0",
     "babel-plugin-module-resolver": "4.0.0",
-    "babel-plugin-react-native-web": "0.14.4",
+    "babel-plugin-react-native-web": "0.14.7",
     "babel-plugin-styled-components": "1.11.1",
-    "eslint": "7.11.0",
-    "eslint-config-prettier": "6.14.0",
+    "eslint": "7.12.1",
+    "eslint-config-prettier": "6.15.0",
     "eslint-plugin-babel": "5.3.1",
     "eslint-plugin-better-styled-components": "1.1.2",
     "eslint-plugin-jest": "24.1.0",
@@ -157,7 +158,7 @@
     "eslint-plugin-sort-destructure-keys": "1.3.5",
     "eslint-plugin-typescript-sort-keys": "1.5.0",
     "husky": "4.3.0",
-    "jest": "26.6.1",
+    "jest": "26.6.3",
     "mockdate": "3.0.2",
     "moment": "2.29.1",
     "prettier": "2.1.2",
@@ -166,17 +167,17 @@
     "react-dom": "16.13.1",
     "react-native": "0.61.5",
     "react-native-gesture-handler": "1.8.0",
-    "react-native-reanimated": "2.0.0-alpha.7",
+    "react-native-reanimated": "2.0.0-alpha.8",
     "react-native-svg": "12.1.0",
     "react-native-typescript-transformer": "1.2.13",
-    "react-native-web": "0.14.4",
-    "react-styleguidist": "11.1.0",
+    "react-native-web": "0.14.7",
+    "react-styleguidist": "11.1.1",
     "react-test-renderer": "16.13.1",
     "stylelint": "13.7.2",
     "stylelint-config-standard": "20.0.0",
     "stylelint-config-styled-components": "0.1.1",
     "stylelint-processor-styled-components": "1.10.0",
-    "typescript": "4.0.3",
+    "typescript": "4.0.5",
     "webpack": "4.44.2"
   },
   "resolutions": {
