@@ -7,24 +7,25 @@ import {
   View,
 } from 'react-native';
 import { RouteProp, useNavigation, useTheme } from '@react-navigation/native';
-import type { StackNavigationProp } from '@react-navigation/stack';
+import { StackNavigationProp } from '@react-navigation/stack';
 import { AppContext } from '../context/AppContext';
 import { NavigationParamsList } from '../types';
 import { streamTheme } from '../utils/streamTheme';
 import { LeftArrow } from '../icons/LeftArrow';
-
+import {
+  LocalAttachmentType,
+  LocalChannelType,
+  LocalCommandType,
+  LocalEventType,
+  LocalMessageType,
+  LocalResponseType,
+  LocalUserType,
+} from '../types';
 import {
   Channel,
-  ChannelList,
   Chat,
-  DeepPartial,
   MessageInput,
   MessageList,
-  OverlayProvider,
-  Streami18n,
-  Theme,
-  Thread,
-  ThreadContextValue,
 } from 'stream-chat-react-native/v2';
 
 export type ChannelScreenProps = {
@@ -67,39 +68,35 @@ const ChannelHeader: React.FC<ChannelHeaderProps> = ({ title }) => {
   );
 };
 
-export const ChannelScreen: React.FC<ChannelScreenProps> = ({ route: {params: {channelId}} }) => {
+export const ChannelScreen: React.FC<ChannelScreenProps> = ({
+  route: {
+    params: { channelId },
+  },
+}) => {
   const { chatClient } = useContext(AppContext);
   const channel = chatClient?.channel('messaging', channelId);
+
   return (
     <SafeAreaView>
-      <Chat client={chatClient} style={streamTheme}>
-        <View style={{ height: '100%' }}>
+      <View style={{ height: '100%' }}>
+        <Chat client={chatClient} style={streamTheme}>
           <ChannelHeader title={'User'} />
-          <Chat client={chatClient}>
+          <View style={{ flexGrow: 1, flexShrink: 1 }}>
             <Channel channel={channel}>
-              <View style={{ flex: 1 }}>
-                <MessageList<
-                  LocalAttachmentType,
-                  LocalChannelType,
-                  LocalCommandType,
-                  LocalEventType,
-                  LocalMessageType,
-                  LocalResponseType,
-                  LocalUserType
-                >
-                  onThreadSelect={(thread) => {
-                    setThread(thread);
-                    if (channel?.id) {
-                      navigation.navigate('Thread', { channelId: channel.id });
-                    }
-                  }}
-                />
-                <MessageInput />
-              </View>
+              <MessageList<
+                LocalAttachmentType,
+                LocalChannelType,
+                LocalCommandType,
+                LocalEventType,
+                LocalMessageType,
+                LocalResponseType,
+                LocalUserType
+              > />
+              <MessageInput />
             </Channel>
-          </Chat>
-        </View>
-      </Chat>
+          </View>
+        </Chat>
+      </View>
     </SafeAreaView>
   );
 };
