@@ -34,6 +34,67 @@ export type ChatContextValue<
 
 export const ChatContext = React.createContext({} as ChatContextValue);
 
+const areEqual = <
+  At extends UnknownType = DefaultAttachmentType,
+  Ch extends UnknownType = DefaultChannelType,
+  Co extends string = DefaultCommandType,
+  Ev extends UnknownType = DefaultEventType,
+  Me extends UnknownType = DefaultMessageType,
+  Re extends UnknownType = DefaultReactionType,
+  Us extends UnknownType = DefaultUserType
+>(
+  prevProps: PropsWithChildren<{
+    value: ChatContextValue<At, Ch, Co, Ev, Me, Re, Us>;
+  }>,
+  nextProps: PropsWithChildren<{
+    value: ChatContextValue<At, Ch, Co, Ev, Me, Re, Us>;
+  }>,
+) => {
+  const {
+    value: {
+      channel: prevChannel,
+      client: prevClient,
+      connectionRecovering: prevConnectionRecovering,
+      isOnline: prevIsOnline,
+    },
+  } = prevProps;
+  const {
+    value: {
+      channel: nextChannel,
+      client: nextClient,
+      connectionRecovering: nextConnectionRecovering,
+      isOnline: nextIsOnline,
+    },
+  } = nextProps;
+
+  const connectionRecoveringEqual =
+    prevConnectionRecovering === nextConnectionRecovering;
+  if (!connectionRecoveringEqual) return false;
+
+  const isOnlineEqual = prevIsOnline === nextIsOnline;
+  if (!isOnlineEqual) return false;
+
+  const channelEqual =
+    !!prevChannel && !!nextChannel && prevChannel.id === nextChannel.id;
+  if (!channelEqual) return false;
+
+  const clientEqual =
+    prevClient.UUID === nextClient.UUID &&
+    Object.keys(prevClient.activeChannels).length ===
+      Object.keys(nextClient.activeChannels).length &&
+    Object.keys(prevClient.listeners).length ===
+      Object.keys(nextClient.listeners).length &&
+    prevClient.mutedChannels.length === nextClient.mutedChannels.length;
+  if (!clientEqual) return false;
+
+  return true;
+};
+
+const ChatProviderMemoized = React.memo(
+  ChatContext.Provider,
+  areEqual,
+) as typeof ChatContext.Provider;
+
 export const ChatProvider = <
   At extends UnknownType = DefaultAttachmentType,
   Ch extends UnknownType = DefaultChannelType,
@@ -48,9 +109,9 @@ export const ChatProvider = <
 }: PropsWithChildren<{
   value: ChatContextValue<At, Ch, Co, Ev, Me, Re, Us>;
 }>) => (
-  <ChatContext.Provider value={(value as unknown) as ChatContextValue}>
+  <ChatProviderMemoized value={(value as unknown) as ChatContextValue}>
     {children}
-  </ChatContext.Provider>
+  </ChatProviderMemoized>
 );
 
 export const useChatContext = <
