@@ -2,19 +2,34 @@ import React, { PropsWithChildren, useContext } from 'react';
 
 import { getDisplayName } from '../utils/getDisplayName';
 
+import type { TouchableOpacityProps } from 'react-native';
 import type { DebouncedFunc } from 'lodash';
 import type {
   ChannelState,
   MessageResponse,
   StreamChat,
   Message as StreamMessage,
-  UserResponse,
 } from 'stream-chat';
 
 import type { SuggestionCommand } from '../suggestionsContext/SuggestionsContext';
 import type { AttachmentProps } from '../../components/Attachment/Attachment';
+import type { AttachmentActionsProps } from '../../components/Attachment/AttachmentActions';
+import type { CardProps } from '../../components/Attachment/Card';
+import type { FileAttachmentProps } from '../../components/Attachment/FileAttachment';
+import type { FileAttachmentGroupProps } from '../../components/Attachment/FileAttachmentGroup';
+import type { FileIconProps } from '../../components/Attachment/FileIcon';
+import type { GalleryProps } from '../../components/Attachment/Gallery';
+import type { MessageProps } from '../../components/Message/Message';
+import type { MessageAvatarProps } from '../../components/Message/MessageSimple/MessageAvatar';
+import type { MessageContentProps } from '../../components/Message/MessageSimple/MessageContent';
+import type { MessageRepliesProps } from '../../components/Message/MessageSimple/MessageReplies';
+import type { MessageStatusProps } from '../../components/Message/MessageSimple/MessageStatus';
 import type { MessageSimpleProps } from '../../components/Message/MessageSimple/MessageSimple';
+import type { MessageTextProps } from '../../components/Message/MessageSimple/MessageTextContainer';
+import type { MarkdownRules } from '../../components/Message/MessageSimple/utils/renderText';
 import type { Message } from '../../components/MessageList/utils/insertDates';
+import type { ReactionListProps } from '../../components/Message/MessageSimple/ReactionList';
+import type { TDateTimeParserInput } from '../../contexts/translationContext/TranslationContext';
 import type {
   DefaultAttachmentType,
   DefaultChannelType,
@@ -25,8 +40,12 @@ import type {
   DefaultUserType,
   UnknownType,
 } from '../../types/types';
+import type { ReactionData } from '../../utils/utils';
 
-export type Alignment = 'right' | 'left';
+export type ActionProps = {
+  reactionsEnabled?: boolean;
+  repliesEnabled?: boolean;
+};
 
 export type GroupType = 'bottom' | 'middle' | 'single' | 'top';
 
@@ -39,7 +58,7 @@ export type MessageWithDates<
   Us extends UnknownType = DefaultUserType
 > = MessageResponse<At, Ch, Co, Me, Re, Us> & {
   groupStyles: GroupType[];
-  readBy: UserResponse<Us>[];
+  readBy: boolean | number;
 };
 
 export const isEditingBoolean = <
@@ -62,20 +81,95 @@ export type MessagesContextValue<
   Me extends UnknownType = DefaultMessageType,
   Re extends UnknownType = DefaultReactionType,
   Us extends UnknownType = DefaultUserType
-> = {
+> = ActionProps & {
+  /**
+   * Custom UI component for attachment.
+   * Defaults to: https://github.com/GetStream/stream-chat-react-native/blob/master/src/components/Attachment/Attachment.tsx
+   */
+  Attachment: React.ComponentType<AttachmentProps<At>>;
+  /**
+   * Custom UI component to display attachment actions. e.g., send, shuffle, cancel in case of giphy
+   * Defaults to https://github.com/GetStream/stream-chat-react-native/blob/master/src/components/Attachment/AttachmentActions.tsx
+   */
+  AttachmentActions: React.ComponentType<AttachmentActionsProps<At>>;
+  /**
+   * Custom UI component for attachment icon for type 'file' attachment.
+   * Defaults to: https://github.com/GetStream/stream-chat-react-native/blob/master/src/components/Attachment/FileIcon.tsx
+   */
+  AttachmentFileIcon: React.ComponentType<FileIconProps>;
+  /**
+   * Custom UI component to display generic media type e.g. giphy, url preview etc
+   * Defaults to https://github.com/GetStream/stream-chat-react-native/blob/master/src/components/Attachment/Card.tsx
+   */
+  Card: React.ComponentType<CardProps<At>>;
   clearEditingState: () => void;
+  /** Should keyboard be dismissed when messaged is touched */
+  dismissKeyboardOnMessageTouch: boolean;
   editing: boolean | Message<At, Ch, Co, Ev, Me, Re, Us>;
-  editMessage: (
-    updatedMessage: StreamMessage<At, Me, Us>,
-  ) => ReturnType<StreamChat<At, Ch, Co, Ev, Me, Re, Us>['updateMessage']>;
-  emojiData: Array<{
-    icon: string;
-    id: string;
-  }>;
+  editMessage: StreamChat<At, Ch, Co, Ev, Me, Re, Us>['updateMessage'];
+  /**
+   * Custom UI component to display File type attachment.
+   * Defaults to https://github.com/GetStream/stream-chat-react-native/blob/master/src/components/Attachment/FileAttachment.tsx
+   */
+  FileAttachment: React.ComponentType<FileAttachmentProps<At>>;
+  /**
+   * Custom UI component to display group of File type attachments or multiple file attachments (in single message).
+   * Defaults to https://github.com/GetStream/stream-chat-react-native/blob/master/src/components/Attachment/FileAttachmentGroup.tsx
+   */
+  FileAttachmentGroup: React.ComponentType<FileAttachmentGroupProps<At>>;
+  /**
+   * Custom UI component to display image attachments.
+   * Defaults to https://github.com/GetStream/stream-chat-react-native/blob/master/src/components/Attachment/Gallery.tsx
+   */
+  Gallery: React.ComponentType<GalleryProps<At>>;
+  /**
+   * Custom UI component to display Giphy image.
+   * Defaults to https://github.com/GetStream/stream-chat-react-native/blob/master/src/components/Attachment/Card.tsx
+   */
+  Giphy: React.ComponentType<CardProps<At>>;
   hasMore: boolean;
   loadingMore: boolean;
   loadMore: DebouncedFunc<() => Promise<void>>;
+  Message: React.ComponentType<MessageProps>;
+  /**
+   * Custom UI component for the avatar next to a message
+   * Defaults to: https://github.com/GetStream/stream-chat-react-native/blob/master/src/components/Message/MessageSimple/MessageAvatar.tsx
+   **/
+  MessageAvatar: React.ComponentType<MessageAvatarProps>;
+  /**
+   * Custom UI component for message content
+   * Defaults to: https://github.com/GetStream/stream-chat-react-native/blob/master/src/components/Message/MessageSimple/MessageContent.tsx
+   */
+  MessageContent: React.ComponentType<
+    MessageContentProps<At, Ch, Co, Ev, Me, Re, Us>
+  >;
+  /**
+   * Custom message replies component
+   * Defaults to: https://github.com/GetStream/stream-chat-react-native/blob/master/src/components/MessageSimple/MessageReplies.tsx
+   */
+  MessageReplies: React.ComponentType<
+    MessageRepliesProps<At, Ch, Co, Ev, Me, Re, Us>
+  >;
   messages: ChannelState<At, Ch, Co, Ev, Me, Re, Us>['messages'];
+  /**
+   * Custom UI component for MessageSimple
+   * Defaults to: https://github.com/GetStream/stream-chat-react-native/blob/master/src/components/Message/MessageSimple/MessageSimple.tsx
+   */
+  MessageSimple: React.ComponentType<
+    MessageSimpleProps<At, Ch, Co, Ev, Me, Re, Us>
+  >;
+  /**
+   * Custom UI component for message status (delivered/read)
+   * Defaults to: https://github.com/GetStream/stream-chat-react-native/blob/master/src/components/Message/MessageSimple/MessageStatus.tsx
+   */
+  MessageStatus: React.ComponentType<
+    MessageStatusProps<At, Ch, Co, Ev, Me, Re, Us>
+  >;
+  /**
+   * Custom UI component to display reaction list.
+   * Defaults to: https://github.com/GetStream/stream-chat-react-native/blob/master/src/components/Reaction/ReactionList.tsx
+   */
+  ReactionList: React.ComponentType<ReactionListProps>;
   removeMessage: (message: {
     id: string;
     parent_id?: StreamMessage<At, Me, Us>['parent_id'];
@@ -85,6 +179,9 @@ export type MessagesContextValue<
   ) => Promise<void>;
   sendMessage: (message: Partial<StreamMessage<At, Me, Us>>) => Promise<void>;
   setEditingState: (message: Message<At, Ch, Co, Ev, Me, Re, Us>) => void;
+  supportedReactions: ReactionData[];
+  /** Whether message text should be rendered before attachments */
+  textBeforeAttachments: boolean;
   updateMessage: (
     updatedMessage: MessageResponse<At, Ch, Co, Me, Re, Us>,
     extraState?: {
@@ -101,11 +198,141 @@ export type MessagesContextValue<
       >['threads'][string];
     },
   ) => void;
-  Attachment?: React.ComponentType<AttachmentProps<At>>;
-  Message?: React.ComponentType<MessageSimpleProps<At, Ch, Co, Ev, Me, Re, Us>>;
+  /**
+   * Custom UI component to display enriched url preview.
+   * Defaults to https://github.com/GetStream/stream-chat-react-native/blob/master/src/components/Attachment/Card.tsx
+   */
+  UrlPreview: React.ComponentType<CardProps<At, Ch, Co, Ev, Me, Re, Us>>;
+  /**
+   * Provide any additional props for `TouchableOpacity` which wraps inner MessageContent component here.
+   * Please check docs for TouchableOpacity for supported props - https://reactnative.dev/docs/touchableopacity#props
+   */
+  additionalTouchableProps?: Omit<TouchableOpacityProps, 'style'>;
+  /**
+   * Custom UI component to override default cover (between Header and Footer) of Card component.
+   * Accepts the same props as Card component.
+   */
+  CardCover?: React.ComponentType<CardProps<At, Ch, Co, Ev, Me, Re, Us>>;
+  /**
+   * Custom UI component to override default Footer of Card component.
+   * Accepts the same props as Card component.
+   */
+  CardFooter?: React.ComponentType<CardProps<At, Ch, Co, Ev, Me, Re, Us>>;
+  /**
+   * Custom UI component to override default header of Card component.
+   * Accepts the same props as Card component.
+   */
+  CardHeader?: React.ComponentType<CardProps<At, Ch, Co, Ev, Me, Re, Us>>;
+  /**
+   * Optional function to custom format the message date
+   */
+  formatDate?: (date: TDateTimeParserInput) => string;
+  /** Object specifying rules defined within simple-markdown https://github.com/Khan/simple-markdown#adding-a-simple-extension */
+  markdownRules?: MarkdownRules;
+  /**
+   * Custom message footer component
+   */
+  MessageFooter?: React.ComponentType<UnknownType & { testID: string }>;
+  /**
+   * Custom message header component
+   */
+  MessageHeader?: React.ComponentType<UnknownType & { testID: string }>;
+  /** Custom UI component for message text */
+  MessageText?: React.ComponentType<
+    MessageTextProps<At, Ch, Co, Ev, Me, Re, Us>
+  >;
 };
 
 export const MessagesContext = React.createContext({} as MessagesContextValue);
+
+const areEqual = <
+  At extends UnknownType = DefaultAttachmentType,
+  Ch extends UnknownType = DefaultChannelType,
+  Co extends string = DefaultCommandType,
+  Ev extends UnknownType = DefaultEventType,
+  Me extends UnknownType = DefaultMessageType,
+  Re extends UnknownType = DefaultReactionType,
+  Us extends UnknownType = DefaultUserType
+>(
+  prevProps: PropsWithChildren<{
+    value: MessagesContextValue<At, Ch, Co, Ev, Me, Re, Us>;
+  }>,
+  nextProps: PropsWithChildren<{
+    value: MessagesContextValue<At, Ch, Co, Ev, Me, Re, Us>;
+  }>,
+) => {
+  const {
+    value: {
+      additionalTouchableProps: prevAdditionalTouchableProps,
+      dismissKeyboardOnMessageTouch: prevDismissKeyboardOnMessageTouch,
+      editing: prevEditing,
+      hasMore: prevHasMore,
+      loadingMore: prevLoadingMore,
+      markdownRules: prevMarkdownRules,
+      messages: prevMessages,
+      supportedReactions: prevSupportedReactions,
+      textBeforeAttachments: prevTextBeforeAttachments,
+    },
+  } = prevProps;
+  const {
+    value: {
+      additionalTouchableProps: nextAdditionalTouchableProps,
+      dismissKeyboardOnMessageTouch: nextDismissKeyboardOnMessageTouch,
+      editing: nextEditing,
+      hasMore: nextHasMore,
+      loadingMore: nextLoadingMore,
+      markdownRules: nextMarkdownRules,
+      messages: nextMessages,
+      supportedReactions: nextSupportedReactions,
+      textBeforeAttachments: nextTextBeforeAttachments,
+    },
+  } = nextProps;
+
+  const dismissKeyboardOnMessageTouchEqual =
+    prevDismissKeyboardOnMessageTouch === nextDismissKeyboardOnMessageTouch;
+  if (!dismissKeyboardOnMessageTouchEqual) return false;
+
+  const hasMoreEqual = prevHasMore === nextHasMore;
+  if (!hasMoreEqual) return false;
+
+  const loadingMoreEqual = prevLoadingMore === nextLoadingMore;
+  if (!loadingMoreEqual) return false;
+
+  const editingEqual = !!prevEditing === !!nextEditing;
+  if (!editingEqual) return false;
+
+  const textBeforeAttachmentsEqual =
+    prevTextBeforeAttachments === nextTextBeforeAttachments;
+  if (!textBeforeAttachmentsEqual) return false;
+
+  const supportedReactionsEqual =
+    prevSupportedReactions.length === nextSupportedReactions.length;
+  if (!supportedReactionsEqual) return false;
+
+  const messagesEqual = prevMessages.length === nextMessages.length;
+  if (!messagesEqual) return false;
+
+  const additionalTouchablePropsEqual =
+    !!prevAdditionalTouchableProps &&
+    !!nextAdditionalTouchableProps &&
+    Object.keys(prevAdditionalTouchableProps).length ===
+      Object.keys(nextAdditionalTouchableProps).length;
+  if (!additionalTouchablePropsEqual) return false;
+
+  const markdownRulesEqual =
+    !!prevMarkdownRules &&
+    !!nextMarkdownRules &&
+    Object.keys(prevMarkdownRules).length ===
+      Object.keys(nextMarkdownRules).length;
+  if (!markdownRulesEqual) return false;
+
+  return true;
+};
+
+const MessagesProviderMemoized = React.memo(
+  MessagesContext.Provider,
+  areEqual,
+) as typeof MessagesContext.Provider;
 
 export const MessagesProvider = <
   At extends UnknownType = DefaultAttachmentType,
@@ -121,9 +348,9 @@ export const MessagesProvider = <
 }: PropsWithChildren<{
   value: MessagesContextValue<At, Ch, Co, Ev, Me, Re, Us>;
 }>) => (
-  <MessagesContext.Provider value={(value as unknown) as MessagesContextValue}>
+  <MessagesProviderMemoized value={(value as unknown) as MessagesContextValue}>
     {children}
-  </MessagesContext.Provider>
+  </MessagesProviderMemoized>
 );
 
 export const useMessagesContext = <
