@@ -40,6 +40,62 @@ export type ThreadContextValue<
 
 export const ThreadContext = React.createContext({} as ThreadContextValue);
 
+const areEqual = <
+  At extends UnknownType = DefaultAttachmentType,
+  Ch extends UnknownType = DefaultChannelType,
+  Co extends string = DefaultCommandType,
+  Ev extends UnknownType = DefaultEventType,
+  Me extends UnknownType = DefaultMessageType,
+  Re extends UnknownType = DefaultReactionType,
+  Us extends UnknownType = DefaultUserType
+>(
+  prevProps: PropsWithChildren<{
+    value: ThreadContextValue<At, Ch, Co, Ev, Me, Re, Us>;
+  }>,
+  nextProps: PropsWithChildren<{
+    value: ThreadContextValue<At, Ch, Co, Ev, Me, Re, Us>;
+  }>,
+) => {
+  const {
+    value: {
+      thread: prevThread,
+      threadHasMore: prevThreadHasMore,
+      threadLoadingMore: prevThreadLoadingMore,
+      threadMessages: prevThreadMessage,
+    },
+  } = prevProps;
+  const {
+    value: {
+      thread: nextThread,
+      threadHasMore: nextThreadHasMore,
+      threadLoadingMore: nextThreadLoadingMore,
+      threadMessages: nextThreadMessages,
+    },
+  } = nextProps;
+
+  const threadHasMoreEqual = prevThreadHasMore === nextThreadHasMore;
+  if (!threadHasMoreEqual) return false;
+
+  const threadLoadingMoreEqual =
+    prevThreadLoadingMore === nextThreadLoadingMore;
+  if (!threadLoadingMoreEqual) return false;
+
+  const threadEqual =
+    !!prevThread && !!nextThread && prevThread.id === nextThread.id;
+  if (!threadEqual) return false;
+
+  const threadMessagesEqual =
+    prevThreadMessage.length === nextThreadMessages.length;
+  if (!threadMessagesEqual) return false;
+
+  return true;
+};
+
+const ThreadProviderMemoized = React.memo(
+  ThreadContext.Provider,
+  areEqual,
+) as typeof ThreadContext.Provider;
+
 export const ThreadProvider = <
   At extends UnknownType = DefaultAttachmentType,
   Ch extends UnknownType = DefaultChannelType,
@@ -54,9 +110,9 @@ export const ThreadProvider = <
 }: PropsWithChildren<{
   value: ThreadContextValue<At, Ch, Co, Ev, Me, Re, Us>;
 }>) => (
-  <ThreadContext.Provider value={(value as unknown) as ThreadContextValue}>
+  <ThreadProviderMemoized value={(value as unknown) as ThreadContextValue}>
     {children}
-  </ThreadContext.Provider>
+  </ThreadProviderMemoized>
 );
 
 export const useThreadContext = <
@@ -79,8 +135,8 @@ export const useThreadContext = <
   >;
 
 /**
- * Typescript currently does not support partial inference so if MessageContentContext
- * typing is desired while using the HOC withMessageContentContextContext the Props for the
+ * Typescript currently does not support partial inference so if ThreadContext
+ * typing is desired while using the HOC withThreadContextContext the Props for the
  * wrapped component must be provided as the first generic.
  */
 export const withThreadContext = <
